pyproject.toml
[tool.poetry]
name = "alhosletmein"
version = "0.1.0"
description = ""
authors = ["Corban Villa <corbanvilla@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^6.2.5"
requests = "^2.26.0"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
